Make all sizes of type `i64`. (#1124)
This has wide-ranging implications for the types of things in the prelude:

* Functions like `replicate` and `iota` now take `i64` arguments.

* The `from_fraction` function now takes `i64`.

* The `to_i32` function has been removed.

Closes #134.

+++ src/Futhark/CodeGen/ImpCode.hs
@@ -366,7 +366,7 @@ bytes = Count
 
 -- | Convert a count of elements into a count of bytes, given the
 -- per-element size.
-withElemType :: Count Elements (TExp Int32) -> PrimType -> Count Bytes (TExp Int64)
+withElemType :: Count Elements (TExp Int64) -> PrimType -> Count Bytes (TExp Int64)
 withElemType (Count e) t =
   bytes $ sExt64 e * isInt64 (LeafExp (SizeOf t) (IntType Int64))
